+++ task-3/DataItem.cs
@@ -5,7 +5,7 @@ public class DataItem
     public string Id { get; set; }
     public byte[] Payload { get; set; }
 
-    public int GetSize()
+    public long GetSize()
     {
         return Payload.Length;
     }
